Rename to components and move getComponent
